% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs-plaatsen.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities}
\format{\describe{
\item{wp_naam, wp_code}{City name and code}
\item{gm_naam, gm_code}{Municipality name and code}
\item{pv_naam, pv_code}{Province name and code}
\item{ld_naam, ld_code}{Country part name and code}
\item{jaar}{Year}
}}
\source{
http://statline.cbs.nl/StatWeb/search/?Q=woonplaatsen&LA=NL
}
\usage{
cities
}
\description{
All cities in The Netherlands with their administrative regions, 2011-2017.
}
\keyword{datasets}
